Remove now obsolete inline elements and styles (#40)
* Remove now obsolete inline elements and styles

* Remove now obsolete inline styles and elements

* Start to replace font-awesome font with CSS inline glyphs

We do only use a very small set of fa icons, so loading the whole font with hundreds of icons is a large overhead. The individual icons can be downloaded as SVG code and freely used under the same license. We'll keep the reference and the class names. This change replaces the "fa-xyz:before" icons from fonts.css seamlessly, and as always after waiting one day for browser cache timeout is required before the "fa-xyz:before" code from fonts.css can be removed.

* Align software docs links