Update RPi download info (#119)
Clarify on which models the 64-bit image can be used, make the download-specific text a list entry, since it can be a little longer, and remove the obsolete text which refers to the Raspberry Pi 1. Since the multiple downloads for Raspberry Pi already contain quite much text, and since there are multiple models with very different performance features, the general text is not added anymore.

Since we don't want to have the large margin that comes with <ul>, we add the list style to this element manually for now.