Update Font Awesome icons
There was a tiny fix in the downlod icon, while I cannot see the difference. The external link icon is currently not used, hence commented for now, but also minified, which was forgotten before. References to the SVG sources on GitHub have been added. Comments in CSS are all removed via minification during deployiment.