SVG style enhancements
Give SVG thumbnails and project images a white background, to align the logos with other SBC images. Also give them a little padding so that the content does not reach the border of the background. For the project frame for this to work, the padding it already has is migrated to a transparent border. Additionally a fixed aspect-ratio is set here, like for the thumbnails, to align SVG images with others.