Split Raspberry Pi downloads by architecture (#241)
And attempt to solve HTML checker errors: Element “html” not allowed in this context.
This shows up for two CSS files which do not contain the string "html" at all. Somehow the wrong parser/checker is applied on wrong file extension, I guess. This error did show up in the past already, which we solved with a filter that time, but then solved itself: https://github.com/validator/validator/issues/1217