Add UEFI live image
In the past we found thought it is necessary to use Clonezilla for some reason, but actually it is not. Clonezilla also does not much more than recreating the initramfs, but in case of initramfs-tools with "most" kernel modules, that should result in exactly the same initramfs image.

It has been basically tested, found functional and useful for some scenarios, so let's offer it.